% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{algo_vol}
\alias{algo_vol}
\title{Algorithm Volatility}
\usage{
algo_vol(trade.value)
}
\arguments{
\item{trade.value}{Vector of a numeric series which represents the total value generated by a
trade strategy computed in daily.}
}
\description{
Algorithm Volatility
}
