% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R, R/utils.R
\name{alpha_stock}
\alias{alpha_stock}
\title{Alpha value}
\usage{
alpha_stock(stock.price, cash.begin, trade.value, rate, trade.days)

alpha_stock(stock.price, cash.begin, trade.value, rate, trade.days)
}
\arguments{
\item{stock.price}{Vector of a stock price series, time level could be 1day,
1hour or 1min.}

\item{cash.begin}{Cash used at begining of a trade.}

\item{trade.value}{Vector of a numeric series which represents the total
value generated by a trade strategy correspond to stock.price.}

\item{rate}{Risk free rate.}

\item{trade.days}{Time span of day.}

\item{stock.price}{Vector of a stock price series, time level could be 1day,
1hour or 1min.}

\item{cash.begin}{Cash used at begining of a trade.}

\item{trade.value}{Vector of a numeric series which represents the total
value generated by a trade strategy correspond to stock.price.}

\item{rate}{Risk free rate.}

\item{trade.days}{Time span of day.}
}
\description{
Alpha value

Alpha value
}
