% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-quantler-method.R
\name{as.quantler}
\alias{as.quantler}
\title{Conversion functions to coerce data objects of arbitrary classes to class
quantler and back, without losing any attributes of the original format.}
\usage{
as.quantler(object)
}
\description{
Conversion functions to coerce data objects of arbitrary classes to class
quantler and back, without losing any attributes of the original format.
}
