% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back_test00.R
\name{backtest00}
\alias{backtest00}
\title{Make judgements of an input quantler obj.}
\usage{
backtest00(trade.info, cash = 0, is.pic = F, is.pic.whole = T,
  plotlevel = "1day", is.tax = T, tax.rate = 0.003, is.simple = F)
}
\arguments{
\item{trade.info}{A data frame contains the signal of trade information
and stock data.}

\item{cash}{Trade money at start.}

\item{is.pic}{Plot results or not.}

\item{is.pic.whole}{Should the pic output be in one piece together or in a
list which contains all segments of different pics, only valied when is.pic
is True.}

\item{plotlevel}{Values in one of "1min", "60mins", "1day", only valied when
is.pic is True.}

\item{is.tax}{Boolean values of True or False, add trading commission or
not.}

\item{tax.rate}{Tax rate per trade, only valide when set is.tax to True.}

\item{is.simple}{Boolean values of True or False, which means if the
judgement should be done in a simple way to run faster, only valied when
is.pic is True.}
}
\value{
A list contains the information of trading results.
}
\description{
Make judgements of an input quantler obj.
}
\details{
Trade.info must be a data frame in the form of:
  data.frame(
   wind_code(character),
   stock.name(character),
   stock.price(numeric),
   stock.time(POSIXCT("yyyy-mm-dd hh:mm")),
   stock.date(as.Date("yyyy-mm-dd")),
   strategy.b(integer in c(-1, 0, 1), 0:"hold", 1:"buy", -1:"sell"),
   strategy.s(c(0,0,1,-1,0,0,....), 0:"hold", 1:"buy", -1:"sell")
   )
}
\examples{
data(stradeinfo_bp)
bt <- backtest(stradeinfo_bp, cash = 100000, is.pic = T, is.pic.whole = T)
plot(bt$pic)
}
