% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R, R/utils.R
\name{bench_vol}
\alias{bench_vol}
\title{Compute Benchmark Volatility}
\usage{
bench_vol(stock.price)

bench_vol(stock.price)
}
\arguments{
\item{stock.price}{a numeric series in the form of vector, which consists of
daily data of stock price}

\item{stock.price}{a numeric series in the form of vector, which consists of
daily data of stock price}
}
\description{
Compute Benchmark Volatility

Compute Benchmark Volatility
}
