% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{beta_stock}
\alias{beta_stock}
\title{Beta value}
\usage{
beta_stock(stock.price, cash.begin, trade.value)
}
\arguments{
\item{stock.price}{Vector of a stock price series, time level could be 1day, 1hour or 1min.}

\item{cash.begin}{Cash used at begining of a trade.}

\item{trade.value}{Vector of a numeric series which represents the total value generated by a
trade strategy correspond to stock.price.}
}
\description{
Beta value
}
