% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bm_return_v}
\alias{bm_return_v}
\title{Benchmark return}
\usage{
bm_return_v(stock.price, cash.begin)
}
\arguments{
\item{stock.price}{Vector of a stock price series, time level could be 1day,
1hour or 1min.}

\item{cash.begin}{Cash used at begining of a trade.}
}
\description{
Benchmark return
}
