% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{info_rate}
\alias{info_rate}
\title{Compute information ratio.}
\usage{
info_rate(stock.price, cash.begin, trade.value, trade.days)
}
\arguments{
\item{stock.price}{Vector of a stock price series, time level could be 1day,
1hour or 1min.}

\item{cash.begin}{Cash used at begining of a trade.}

\item{trade.value}{Vector of a numeric series which represents the total
value generated by a trade strategy.}

\item{trade.days}{Time span of day.}
}
\description{
Compute information ratio.
}
