% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{maxdrawdown}
\alias{maxdrawdown}
\title{Compute max draw down of a input stock price series}
\usage{
maxdrawdown(trade.value)
}
\arguments{
\item{trade.value}{Vector of a numeric series which represents the total
value generated by a trade strategy.}
}
\value{
a list with two elements of 'value' and 'loc'.
}
\description{
\code{maxdrawdown} returns a list with two elements (value ,loc ), which represents
   value of max draw down and its location.
}
\examples{
data(sh600000)
maxdd <- maxdrawdown(sh600000$close/10000)

}
