% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fun.R
\docType{methods}
\name{plot,quantler,missing-method}
\alias{plot,quantler,missing-method}
\title{Plot a quantler object}
\usage{
\S4method{plot}{quantler,missing}(x, ..., plotlevel = "1day",
  is.simple = F, segments = F)
}
\arguments{
\item{x}{A quantler object}

\item{plotlevel}{A character variable must be one of c('1day', '60mins',
'1min')}

\item{segments}{A logical variable. When segments is TRUE this function will
return a list contians all the sub pictures, and when FALSE plot all the
segments together.}
}
\value{
A picture or a list contains all the segments of the picture.
}
\description{
A generic function of class quantler.
}
\examples{
   data(stradeinfo_bp)
   test.qtler <- backtest(stradeinfo_bp, cash = 100000, is.tax=T, tax.rate = 0.002)
   plot(test.qtler)
}
