% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_pre.R
\name{pre_correct}
\alias{pre_correct}
\title{Calculate correct rate of buy and sell signal, based on stock price}
\usage{
pre_correct(stock.price = NA, strategy.b = NA, strategy.s = NA)
}
\arguments{
\item{stock.price}{a one dimension vector of numeric data, which represents a
time series of stock price.}

\item{strategy.b}{the trade signal of buy operation consists of a one
dimension vector contians integer values of -1,0,1, of which '-1' equals to
'sell','0' equals to 'hold' and '1' equals to 'buy'.}

\item{strategy.s}{the trade signal of sell operation consists of a one
dimension vector contians integer values of -1,0,1, of which '-1' equals to
'sell','0' equals to 'hold' and '1' equals to 'buy'.}
}
\value{
a vector with two values represent
}
\description{
Calculate correct rate of buy and sell signal, based on stock price
}
