% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-def.R
\docType{class}
\name{quantler-class}
\alias{quantler-class}
\alias{quantler}
\title{An S4 class for quant strategy judgements.}
\description{
An S4 class for quant strategy judgements.
}
\section{Slots}{

\describe{
\item{\code{trade.info}}{A data.frame}

\item{\code{data.level}}{A character value of one of c('1day', '60mins', '1min')}

\item{\code{bt.features}}{A list contains severial features about quant strategy
judegments produced by the back-test process.}
}}

