% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sharpe_rate}
\alias{sharpe_rate}
\title{Compute charpe ratio of the input trade.value}
\usage{
sharpe_rate(trade.value, rate, trade.days)
}
\arguments{
\item{trade.value}{vector of a numeric series which represents the total
value generated by a trade strategy.}

\item{rate}{risk free rate}

\item{trade.days}{time span of day}
}
\description{
Compute charpe ratio of the input trade.value
}
