% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-quantler-method.R
\name{is.quantler}
\alias{is.quantler}
\title{Is an object a quantler object?}
\usage{
is.quantler(object)
}
\arguments{
\item{object}{An object.}
}
\value{
A logical value. TRUE when the input is a quantler object and FALSE
  otherwise.
}
\description{
Checks for quantler objects.
}
