% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-def.R
\docType{class}
\name{quantler-class}
\alias{quantler-class}
\alias{quantler}
\title{An S4 class to represent a trade data set producted by quant model}
\description{
An S4 class to represent a trade data set producted by quant model
}
\section{Slots}{

\describe{
\item{\code{trade.info}}{A data.frame}
}}

